refactor: reorganize code and introduce broadcast bin

+++ crates/oc_utils/src/d2.rs
@@ -1,7 +1,9 @@
 use oc_root::WORLD_WIDTH;
+use rkyv::{Archive, Deserialize, Serialize};
 
-#[derive(Debug, Clone, Copy)]
-pub struct Xy(pub usize, pub usize);
+#[derive(Debug, Clone, Copy, Archive, Deserialize, Serialize, PartialEq)]
+#[rkyv(compare(PartialEq), derive(Debug))]
+pub struct Xy(pub u64, pub u64);
 
 #[derive(Debug, Clone, Copy)]
 pub struct XyIndex(pub usize);
@@ -10,17 +12,17 @@ impl From<XyIndex> for Xy {
     fn from(XyIndex(i): XyIndex) -> Self {
         let x = i % WORLD_WIDTH;
         let y = i / WORLD_WIDTH;
-        Self(x, y)
+        Self(x as u64, y as u64)
     }
 }
 
 impl From<Xy> for XyIndex {
     fn from(Xy(x, y): Xy) -> Self {
-        Self(y * WORLD_WIDTH + x)
+        Self(y as usize * WORLD_WIDTH + x as usize)
     }
 }
 
-impl From<Xy> for (usize, usize) {
+impl From<Xy> for (u64, u64) {
     fn from(value: Xy) -> Self {
         (value.0, value.1)
     }
